# andes-api-toolkit/contracts/openapi-server.yaml
openapi: 3.0.3
info:
  title: Andes Customer API
  description: API propia expuesta por el PoC Server. Gestiona clientes (customers).
  version: 1.0.0
  contact:
    name: Andes API Toolkit Team
    email: support@andes.pe
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
tags:
  - name: Customers
    description: Operations about customers
paths:
  /api/v1/customers:
    get:
      operationId: listCustomers
      tags: [Customers]
      summary: List customers (paginated)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Page of customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPageEnvelope'
    post:
      operationId: createCustomer
      tags: [Customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
      responses:
        '201':
          description: Created customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerEnvelope'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/customers/{id}:
    get:
      operationId: getCustomerById
      tags: [Customers]
      summary: Get a customer by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerEnvelope'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: deleteCustomer
      tags: [Customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Deleted
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  schemas:
    Customer:
      type: object
      properties:
        id: { type: integer, format: int64 }
        fullName: { type: string }
        email: { type: string, format: email }
        createdAt: { type: string, format: date-time }
    CustomerRequest:
      type: object
      required: [fullName, email]
      properties:
        fullName: { type: string, minLength: 3 }
        email: { type: string, format: email }
    CustomerEnvelope:
      type: object
      description: Standard Andes ApiResponse envelope wrapping a single Customer
      properties:
        success: { type: boolean }
        data:
          $ref: '#/components/schemas/Customer'
        error:
          nullable: true
          $ref: '#/components/schemas/ApiError'
        metadata:
          $ref: '#/components/schemas/ApiMetadata'
    CustomerPageEnvelope:
      type: object
      properties:
        success: { type: boolean }
        data:
          $ref: '#/components/schemas/CustomerPage'
        error:
          nullable: true
          $ref: '#/components/schemas/ApiError'
        metadata:
          $ref: '#/components/schemas/ApiMetadata'
    CustomerPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        data: { nullable: true }
        error:
          $ref: '#/components/schemas/ApiError'
        metadata:
          $ref: '#/components/schemas/ApiMetadata'
    ApiError:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        httpStatus: { type: integer }
        traceId: { type: string }
        timestamp: { type: string, format: date-time }
        details:
          type: array
          items:
            $ref: '#/components/schemas/ApiErrorDetail'
    ApiErrorDetail:
      type: object
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
        rejectedValue: { nullable: true }
    ApiMetadata:
      type: object
      properties:
        traceId: { type: string }
        correlationId: { type: string }
        requestId: { type: string }
        apiVersion: { type: string }
        timestamp: { type: string, format: date-time }
    Pagination:
      type: object
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT